Moved all ui api endpoints to api/ui
